+++ src/components/MainPage.jsx
@@ -1,9 +1,0 @@
-import React from 'react';
-
-function Main () {
-    return (
-        <div className='MainPage'>Main</div>
-    );
-}
-
-export default Main;