+++ src/data/data.js
@@ -1,21 +1,126 @@
 export const data = [
     {
-        question: "Are you 100% real?",
-        points: 0
+        question: "I see myself as someone who is reserved.",
+        cat: 'E',
+        supports: -1,
+        type: 1
     },
     {
-        question: "Are you 100% gangster?",
-        points: 0
+        question: "I see myself as someone who is generally trusting.",
+        cat: 'A',
+        supports: 1,
+        type: 1
     },
     {
-        question: "Would you say you are Carmen Winstead",
-        points: 0
+        question: "I see myself as someone who tends to be lazy.",
+        cat: 'N',
+        supports: 1,
+        type: 1
     },
     {
-        question: "Would you say you are goated with the sauce?",
-        points: 0
+        question: "I see myself as someone who is relaxed, handles stress really well.",
+        cat: 'C',
+        supports: 1,
+        type: 1
     },
     {
-        filler: "filler"
+        question: "I see myself as someone who has few artistic interest.",
+        cat: 'O',
+        supports: 0,
+        type: 1
+    },
+    {
+        question: "I see myself as someone who is outgoing, sociable.",
+        cat: 'E',
+        supports: 1,
+        type: 1
+    },
+    {
+        question: "I see myself as someone who tends to find fault with others.",
+        cat: 'A',
+        supports: 0,
+        type: 1
+    },
+    {
+        question: "I see myself as someone who does a thorough job.",
+        cat: 'C',
+        supports: 1,
+        type: 1
+    },
+    {
+        question: "I see myself as someone who gets nervous easily.",
+        cat: 'N',
+        supports: 1,
+        type: 1
+    },
+    {
+        question: "I see myself as someone who has an active imagination.",
+        cat: 'O',
+        supports: 1,
+        type: 1
+    },
+    {
+        question: "I see myself as someone who is even tempered.",
+        cat: 'N',
+        supports: 0,
+        type: 1
+    },
+    {
+        question: "I see myself as someone who rarely feels lonely",
+        cat: 'N',
+        supports: 0,
+        type: 1
+    },
+    {
+        question: "I see myself as someone who often has been the leader of a group.",
+        cat: 'E',
+        supports: 1,
+        type: 1
+    },
+    {
+        question: "I see myself as someone who is an active person.",
+        cat: 'E',
+        supports: 1,
+        type: 1
+    },
+    {
+        question: "I see myself as someone who likes to daydream",
+        cat: 'O',
+        supports: 1,
+        type: 1
+    },
+    {
+        question: "I see myself as someone who prefers familiar surroundings.",
+        tcat: 'O',
+        supports: 0,
+        type: 1
+    },
+    {
+        question: "I see myself as someone who is more aware of the human side.",
+        cat: 'A',
+        supports: 1,
+        type: 1
+    },
+    {
+        question: "I see myself as someone who is sarcastic and cutting.",
+        cat: 'A',
+        supports: 0,
+        type: 1
+    },
+    {
+        question: "I see myself as someone who makes intelligent decisions.",
+        cat: 'C',
+        supports: 1,
+        type: 1
+    },
+    {
+        question: "I see myself as someone who has a clear set of goals",
+        cat: 'C',
+        supports: 1,
+        type: 1
+    },
+    {
+        filler: "I see myself as someone who is epic",
+        type:-1
     }
 ]