+++ CHANGELOG.md
@@ -1,4 +1,15 @@
 All notable changes to this project will be documented in this file.
+## 0.2.3 - 2022-04-02
+### Added
+- Literally the entire results page and most if not all the testing functionality
+- The Assessment class, which essentially holds all assessment information until results are either saved to the backend or discarded.
+### Deprecated
+- The "card" class for anything other than bootstrap cards
+### Removed
+- A good chunk of previous css and html for Carousel and CustomCard components
+### Fixed
+- The assessment actually does something now using SessionStorage.
+
 ## 0.2.1 - 2022-03-26
 ### Added
 - README and CHANGELOG info